refactor(wp): move getPosts to Post model
